requirements.txt
aiohttp==4.0.0a1
aioredis==1.3.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
configparser==5.2.0
cryptography==37.0.3
dependency-injector==4.38.0
Deprecated==1.2.13
fastapi==0.73.0
fastapi-cache2==0.1.8
fastapi-utils==0.2.1
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httplib2==0.20.2
idna==3.3
loguru==0.6.0
msgpack==1.0.3
multidict==4.7.6
Naked==0.1.31
oauth2client==4.1.3
packaging==21.3
pandas==1.5.3
pendulum==2.1.2
pika==1.3.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pydash==6.0.2
pydantic==1.9.0
PyJWT==2.3.0
pymongo==4.1.1
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
pytzdata==2020.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
supervisor==4.2.4
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.4
wrapt==1.13.3
yarl==1.7.2
